Refactor: move configurations to new 'config' package
