+++ internal/validator/config/validation_configuration.go
@@ -1,4 +1,4 @@
-package validator
+package config
 
 import "time"
