+++ internal/validator/report_nodes.go
@@ -2,6 +2,7 @@ package validator
 
 import (
 	"github.com/aml-org/amf-custom-validator/internal/types"
+	"github.com/aml-org/amf-custom-validator/internal/validator/config"
 	"time"
 )
 
@@ -16,7 +17,7 @@ func DialectInstance(report *types.ObjectMap, context *types.ObjectMap) []types.
 	return []types.ObjectMap{dialectInstance}
 }
 
-func ValidationReportNode(profileName string, results []any, conforms bool, validationConfig ValidationConfiguration, reportConfig ReportConfiguration) types.ObjectMap {
+func ValidationReportNode(profileName string, results []any, conforms bool, validationConfig config.ValidationConfiguration, reportConfig config.ReportConfiguration) types.ObjectMap {
 	reportTypes := []string{"reportSchema:ReportNode", "shacl:ValidationReport"}
 	report := types.ObjectMap{
 		"@id":         "validation-report",
