+++ internal/validator/test_utils.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/aml-org/amf-custom-validator/internal/config"
 	p "github.com/aml-org/amf-custom-validator/internal/parser/profile"
+	config2 "github.com/aml-org/amf-custom-validator/internal/validator/config"
 	"github.com/aml-org/amf-custom-validator/pkg/milestones"
 	"io/ioutil"
 	"strings"
@@ -28,7 +29,7 @@ func validate(profile relativePath, data relativePath) string {
 	p.GenReset()
 	profileText := read(profile)
 	dataText := read(data)
-	report, err := ValidateWithConfiguration(profileText, dataText, config.Debug, nil, TestValidationConfiguration{}, DefaultReportConfiguration())
+	report, err := ValidateWithConfiguration(profileText, dataText, config.Debug, nil, config2.TestValidationConfiguration{}, config2.DefaultReportConfiguration())
 	if err != nil {
 		panic(err)
 	}
