+++ internal/validator/validate_configuration_test.go
@@ -3,18 +3,19 @@ package validator
 import (
 	"encoding/json"
 	"github.com/aml-org/amf-custom-validator/internal/config"
+	config2 "github.com/aml-org/amf-custom-validator/internal/validator/config"
 	"testing"
 )
 
 func TestNoDateCreated(t *testing.T) {
-	reportConfig := ReportConfiguration{
+	reportConfig := config2.ReportConfiguration{
 		IncludeReportCreationTime: false,
 	}
 
 	profile := read("../../test/data/integration/profile1/profile.yaml")
 	data := read("../../test/data/integration/profile1/negative.data.jsonld")
 
-	report, err := ValidateWithConfiguration(profile, data, config.Debug, nil, TestValidationConfiguration{}, reportConfig)
+	report, err := ValidateWithConfiguration(profile, data, config.Debug, nil, config2.TestValidationConfiguration{}, reportConfig)
 
 	if err != nil {
 		t.Errorf("Error during validation\n")
